+++ collector/src/db/models/mesh.rs
@@ -1,5 +1,5 @@
 pub struct Mesh {
-	pub id: Option<i64>,
+	pub id: Option<u32>,
 	pub name: String,
 	pub description: String,
 	pub enabled: bool,
